+++ pkg/tests/cross-tests/assert.go
@@ -1,0 +1,30 @@
+package crosstests
+
+import (
+	"github.com/hashicorp/go-cty/cty"
+	"github.com/stretchr/testify/require"
+)
+
+func FailNotEqual(t T, name string, tfVal, pulVal any) {
+	t.Logf(name + " not equal!")
+	t.Logf("TF value %s", tfVal)
+	t.Logf("PU value %s", pulVal)
+	t.Fail()
+}
+
+func assertCtyValEqual(t T, name string, tfVal, pulVal cty.Value) {
+	if !tfVal.RawEquals(pulVal) {
+		FailNotEqual(t, name, tfVal.GoString(), pulVal.GoString())
+	}
+}
+
+func assertValEqual(t T, name string, tfVal, pulVal any) {
+	// usually plugin-sdk schema types
+	if hasEqualTfVal, ok := tfVal.(interface{ Equal(interface{}) bool }); ok {
+		if !hasEqualTfVal.Equal(pulVal) {
+			FailNotEqual(t, name, tfVal, pulVal)
+		}
+	} else {
+		require.Equal(t, tfVal, pulVal, "Values for key %s do not match", name)
+	}
+}
