+++ pkg/tests/cross-tests/defaults.go
@@ -1,0 +1,9 @@
+package crosstests
+
+var (
+	defProviderShortName = "crossprovider"
+	defRtype             = "crossprovider_testres"
+	defRtok              = "TestRes"
+	defRtoken            = defProviderShortName + ":index:" + defRtok
+	DefProviderVer       = "0.0.1"
+)
