+++ pkg/tests/cross-tests/input_check.go
@@ -15,12 +15,10 @@ package crosstests
 import (
 	"context"
 
-	"github.com/hashicorp/go-cty/cty"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tests/pulcheck"
-	"github.com/stretchr/testify/require"
 )
 
 // Adapted from diff_check.go
@@ -34,64 +32,6 @@ type inputTestCase struct {
 	SkipCompareRawPlan   bool
 	SkipCompareRawConfig bool
 	SkipCompareRawState  bool
-}
-
-func FailNotEqual(t T, name string, tfVal, pulVal any) {
-	t.Logf(name + " not equal!")
-	t.Logf("TF value %s", tfVal)
-	t.Logf("PU value %s", pulVal)
-	t.Fail()
-}
-
-func assertCtyValEqual(t T, name string, tfVal, pulVal cty.Value) {
-	if !tfVal.RawEquals(pulVal) {
-		FailNotEqual(t, name, tfVal.GoString(), pulVal.GoString())
-	}
-}
-
-func assertValEqual(t T, name string, tfVal, pulVal any) {
-	// usually plugin-sdk schema types
-	if hasEqualTfVal, ok := tfVal.(interface{ Equal(interface{}) bool }); ok {
-		if !hasEqualTfVal.Equal(pulVal) {
-			FailNotEqual(t, name, tfVal, pulVal)
-		}
-	} else {
-		require.Equal(t, tfVal, pulVal, "Values for key %s do not match", name)
-	}
-}
-
-func ensureProviderValid(t T, tfp *schema.Provider) {
-	for _, r := range tfp.ResourcesMap {
-		//nolint:staticcheck
-		if r.Read == nil && r.ReadContext == nil {
-			r.ReadContext = func(_ context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-				return nil
-			}
-		}
-		if r.DeleteContext == nil {
-			r.DeleteContext = func(
-				ctx context.Context, rd *schema.ResourceData, i interface{},
-			) diag.Diagnostics {
-				return diag.Diagnostics{}
-			}
-		}
-
-		if r.CreateContext == nil {
-			r.CreateContext = func(
-				ctx context.Context, rd *schema.ResourceData, i interface{},
-			) diag.Diagnostics {
-				rd.SetId("newid")
-				return diag.Diagnostics{}
-			}
-		}
-
-		r.UpdateContext = func(
-			ctx context.Context, rd *schema.ResourceData, i interface{},
-		) diag.Diagnostics {
-			return diag.Diagnostics{}
-		}
-	}
-	require.NoError(t, tfp.InternalValidate())
 }
 
 // Adapted from diff_check.go
@@ -112,27 +52,21 @@ func runCreateInputCheck(t T, tc inputTestCase) {
 		rd.SetId("someid") // CreateContext must pick an ID
 		return make(diag.Diagnostics, 0)
 	}
-	var (
-		providerShortName = "crossprovider"
-		rtype             = "crossprovider_testres"
-		rtok              = "TestRes"
-		rtoken            = providerShortName + ":index:" + rtok
-	)
 
 	tfwd := t.TempDir()
 
-	tfd := newTfDriver(t, tfwd, providerShortName, rtype, tc.Resource)
-	tfd.writePlanApply(t, tc.Resource.Schema, rtype, "example", tc.Config)
+	tfd := newTfDriver(t, tfwd, defProviderShortName, defRtype, tc.Resource)
+	tfd.writePlanApply(t, tc.Resource.Schema, defRtype, "example", tc.Config)
 
-	resMap := map[string]*schema.Resource{rtype: tc.Resource}
-	bridgedProvider := pulcheck.BridgedProvider(t, providerShortName, resMap)
+	resMap := map[string]*schema.Resource{defRtype: tc.Resource}
+	bridgedProvider := pulcheck.BridgedProvider(t, defProviderShortName, resMap)
 	pd := &pulumiDriver{
-		name:                providerShortName,
-		pulumiResourceToken: rtoken,
-		tfResourceName:      rtype,
+		name:                defProviderShortName,
+		pulumiResourceToken: defRtoken,
+		tfResourceName:      defRtype,
 		objectType:          tc.ObjectType,
 	}
-	yamlProgram := pd.generateYAML(t, bridgedProvider.P.ResourcesMap() ,tc.Config)
+	yamlProgram := pd.generateYAML(t, bridgedProvider.P.ResourcesMap(), tc.Config)
 
 	pt := pulcheck.PulCheck(t, bridgedProvider, string(yamlProgram))
